Add files via upload
About the working of a 3D printer, check out the online resources. About what's an STL file, and then learn to slice it (using ultimaker or creality slicer).About bed temperature, infill density and other printer settings. Finally get an STL file from the internet, and slice it and put it for print.

Resources:

● Ultimaker Cura

● PLA settings

● How printer works

● autodesk fusion 360